Move function to utils folder
