+++ packages/core/content-type-builder/admin/src/components/FormModal/forms/utils/getUsedAttributeNames.js
@@ -1,0 +1,15 @@
+/**
+ *
+ * @param {array} attributes The attributes found on the dataManager's modifiedData object
+ * @param {object} schemaData The modifiedData and SchemaData objects from the reducer state
+ * @returns A list of names already being used
+ */
+const getUsedAttributeNames = (attributes, schemaData) => {
+  return attributes
+    .filter(({ name }) => {
+      return name !== schemaData.initialData.name;
+    })
+    .map(({ name }) => name);
+};
+
+export default getUsedAttributeNames;
